requirements.txt
asgiref==3.6.0
astroid==3.0.2
awscli==1.32.14
boto3==1.34.14
botocore==1.34.14
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
colorama==0.4.4
crispy-bootstrap4==2022.1
cryptography==40.0.2
defusedxml==0.7.1
dill==0.3.7
distlib==0.3.6
dj-database-url==0.5.0
dj3-cloudinary-storage==0.0.6
Django==3.2
django-allauth==0.41.0
django-countries==7.2.1
django-crispy-forms==2.0
django-js-asset==2.0.0
django-mptt==0.14.0
django-paypal==2.0
django-phone-field==1.8.1
django-storages==1.13.2
docutils==0.16
filelock==3.12.0
gunicorn==20.1.0
idna==3.4
isort==5.13.2
jmespath==1.0.1
mccabe==0.7.0
oauthlib==3.2.2
paypal-checkout-serversdk==1.0.1
paypalhttp==1.0.1
paypalrestsdk==1.13.0
Pillow==9.4.0
platformdirs==3.2.0
psycopg2-binary==2.9.6
pyasn1==0.5.1
pycodestyle==2.10.0
pycparser==2.21
pylint==3.0.3
pylint-django==2.5.5
pylint-plugin-utils==0.8.2
pyOpenSSL==23.1.1
python-dateutil==2.8.2
python-decouple==3.8
python3-openid==3.2.0
pytz==2022.7.1
PyYAML==6.0.1
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.7.2
s3transfer==0.10.0
simplejson==3.19.1
six==1.16.0
sqlparse==0.4.3
stripe==5.2.0
tomlkit==0.12.3
tzlocal==5.2
urllib3==1.26.15
virtualenv==20.21.0
